requirements.txt
fastapi
uvicorn
paho-mqtt
pyserial
pynmeagps
